factor out datasource_access_request model (#8809)
