+++ superset/models/core.py
@@ -1322,74 +1322,6 @@ class FavStar(Model):  # pylint: disable=too-few-public-methods
     dttm = Column(DateTime, default=datetime.utcnow)
 
 
-class DatasourceAccessRequest(Model, AuditMixinNullable):
-    """ORM model for the access requests for datasources and dbs."""
-
-    __tablename__ = "access_request"
-    id = Column(Integer, primary_key=True)  # pylint: disable=invalid-name
-
-    datasource_id = Column(Integer)
-    datasource_type = Column(String(200))
-
-    ROLES_BLACKLIST = set(config["ROBOT_PERMISSION_ROLES"])
-
-    @property
-    def cls_model(self) -> Type["BaseDatasource"]:
-        return ConnectorRegistry.sources[self.datasource_type]
-
-    @property
-    def username(self) -> Markup:
-        return self.creator()
-
-    @property
-    def datasource(self) -> "BaseDatasource":
-        return self.get_datasource
-
-    @datasource.getter  # type: ignore
-    @utils.memoized
-    def get_datasource(self) -> "BaseDatasource":
-        ds = db.session.query(self.cls_model).filter_by(id=self.datasource_id).first()
-        return ds
-
-    @property
-    def datasource_link(self) -> Optional[Markup]:
-        return self.datasource.link  # pylint: disable=no-member
-
-    @property
-    def roles_with_datasource(self) -> str:
-        action_list = ""
-        perm = self.datasource.perm  # pylint: disable=no-member
-        pv = security_manager.find_permission_view_menu("datasource_access", perm)
-        for role in pv.role:
-            if role.name in self.ROLES_BLACKLIST:
-                continue
-            # pylint: disable=no-member
-            href = (
-                f"/superset/approve?datasource_type={self.datasource_type}&"
-                f"datasource_id={self.datasource_id}&"
-                f"created_by={self.created_by.username}&role_to_grant={role.name}"
-            )
-            link = '<a href="{}">Grant {} Role</a>'.format(href, role.name)
-            action_list = action_list + "<li>" + link + "</li>"
-        return "<ul>" + action_list + "</ul>"
-
-    @property
-    def user_roles(self) -> str:
-        action_list = ""
-        for role in self.created_by.roles:  # pylint: disable=no-member
-            # pylint: disable=no-member
-            href = (
-                f"/superset/approve?datasource_type={self.datasource_type}&"
-                f"datasource_id={self.datasource_id}&"
-                f"created_by={self.created_by.username}&role_to_extend={role.name}"
-            )
-            link = '<a href="{}">Extend {} Role</a>'.format(href, role.name)
-            if role.name in self.ROLES_BLACKLIST:
-                link = "{} Role".format(role.name)
-            action_list = action_list + "<li>" + link + "</li>"
-        return "<ul>" + action_list + "</ul>"
-
-
 # events for updating tags
 if is_feature_enabled("TAGGING_SYSTEM"):
     sqla.event.listen(Slice, "after_insert", ChartUpdater.after_insert)
