+++ tests/core_tests.py
@@ -40,6 +40,7 @@ from superset.connectors.sqla.models import SqlaTable
 from superset.db_engine_specs.base import BaseEngineSpec
 from superset.db_engine_specs.mssql import MssqlEngineSpec
 from superset.models import core as models
+from superset.models.datasource_access_request import DatasourceAccessRequest
 from superset.models.sql_lab import Query
 from superset.utils import core as utils
 from superset.views import core as views
@@ -55,7 +56,7 @@ class CoreTests(SupersetTestCase):
 
     def setUp(self):
         db.session.query(Query).delete()
-        db.session.query(models.DatasourceAccessRequest).delete()
+        db.session.query(DatasourceAccessRequest).delete()
         db.session.query(models.Log).delete()
         self.table_ids = {
             tbl.table_name: tbl.id for tbl in (db.session.query(SqlaTable).all())
